+++ runtime/core/Clownfish/Obj.c
@@ -68,7 +68,7 @@ Obj_To_String_IMP(Obj *self) {
 #elif (CHY_SIZEOF_PTR == 8)
     int64_t   iaddress   = CHY_PTR_TO_I64(self);
     uint64_t  address    = (uint64_t)iaddress;
-    uint32_t  address_hi = address >> 32;
+    uint32_t  address_hi = (uint32_t)(address >> 32);
     uint32_t  address_lo = address & 0xFFFFFFFF;
     return Str_newf("%o@0x%x32%x32", Obj_get_class_name(self), address_hi,
                     address_lo);
